refactor(chatpanel): split ChatPanelTabBar into a module directory
Convert ChatPanelTabBar.tsx into ChatPanelTabBar/ with index.tsx plus
TabPill, TabPillHoverCard, tabPillHoverCardData and ChatPanelPlusMenu;
the test moves alongside. index.tsx re-exports every previous named
export, so "./ChatPanelTabBar" still resolves for both importers.
